Read system information during jmhrun in order to ensure that running jmhplot on a different system doesn't impact the result
